requirements.txt
feedparser>=6.0
google-genai>=1.0.0
